+++ lib/whois/record/parser/whois.nic.tv.rb
@@ -7,8 +7,7 @@
 #++
 
 
-require 'whois/record/parser/base'
-require 'whois/record/scanners/verisign'
+require 'whois/record/parser/base_verisign'
 
 
 module Whois
@@ -16,77 +15,11 @@ module Whois
     class Parser
 
       # Parser for the whois.nic.tv server.
-      class WhoisNicTv < Base
-        include Scanners::Scannable
-
-        self.scanner = Scanners::Verisign
-
-
-        property_supported :disclaimer do
-          node("Disclaimer")
-        end
-
-
-        property_supported :domain do
-          node("Domain Name", &:downcase)
-        end
-
-        property_supported :domain_id do
-          node("Domain ID", &:downcase)
-        end
-
-
-        property_supported :status do
-          # node("Status")
-          if available?
-            :available
-          else
-            :registered
-          end
-        end
-
-        property_supported :available? do
-          !!(content_for_scanner =~ /^No match for/)
-        end
-
-        property_supported :registered? do
-          !available?
-        end
-
-
-        property_supported :created_on do
-          node("Creation Date") { |raw| Time.parse(raw) }
-        end
-
-        property_supported :updated_on do
-          node("Updated Date") { |raw| Time.parse(raw) }
-        end
-
-        property_supported :expires_on do
-          node("Expiration Date") { |raw| Time.parse(raw) }
-        end
-
-
-        property_supported :registrar do
-          nil
-        end
-
-
-        property_supported :nameservers do
-          Array.wrap(node("Name Server")).reject { |value| value =~ /no nameserver/i }.map do |name|
-            Record::Nameserver.new(name: name.downcase)
-          end
-        end
-
-
-        def referral_whois
-          node("Whois Server")
-        end
-
-        def referral_url
-          node("Referral URL")
-        end
-
+      #
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
+      #
+      class WhoisNicTv < BaseVerisign
       end
 
     end