+++ spec/whois/record/parser/responses/whois.crsnic.net/com/property_registrar_with_multiple_entries_spec.rb
@@ -26,7 +26,7 @@ describe Whois::Record::Parser::WhoisCrsnicNet, "property_registrar_with_multipl
       expect(subject.registrar).to be_a(Whois::Record::Registrar)
       expect(subject.registrar.id).to eq(nil)
       expect(subject.registrar.name).to eq("MARKMONITOR INC.")
-      expect(subject.registrar.organization).to eq("MARKMONITOR INC.")
+      expect(subject.registrar.organization).to eq(nil)
       expect(subject.registrar.url).to eq("http://www.markmonitor.com")
     end
   end