+++ spec/whois/record/parser/responses/whois.nic.cc/status_available_spec.rb
@@ -38,7 +38,7 @@ describe Whois::Record::Parser::WhoisNicCc, "status_available.expected" do
   end
   describe "#status" do
     it do
-      expect(subject.status).to eq(nil)
+      expect(subject.status).to eq(:available)
     end
   end
   describe "#available?" do